Trying to generify teleporting
