+++ src/main/java/me/gt3ch1/betterlife/commands/HOME.java
@@ -5,6 +5,7 @@ import java.util.LinkedHashMap;
 import me.gt3ch1.betterlife.Main.Main;
 import me.gt3ch1.betterlife.commandhelpers.BetterLifeCommands;
 import me.gt3ch1.betterlife.data.BL_HOME;
+import me.gt3ch1.betterlife.eventhelpers.PlayerTeleportHelper;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.command.Command;
@@ -18,10 +19,10 @@ import org.bukkit.scheduler.BukkitTask;
  * This class represents the /home command for BetterLife.
  */
 public class HOME extends BetterLifeCommands implements CommandExecutor {
-    private int checkerID;
-    private int taskID;
+
     BL_HOME homeGetter = Main.bl_home;
 
+    private PlayerTeleportHelper tpHelper = new PlayerTeleportHelper();
     /**
      * Initializes the home command.
      *
@@ -63,7 +64,7 @@ public class HOME extends BetterLifeCommands implements CommandExecutor {
             case 0:
                 if (homeList.size() == 1) {
                     home = homeList.get(listOfHomes[0]);
-                    teleportPlayer(player, home, listOfHomes[0]);
+                    tpHelper.teleportPlayer(player, home, listOfHomes[0]);
                 } else if (homeList.size() == 0) {
                     sendMessage(player, "&4You don't have any homes!", true);
                     return false;
@@ -83,7 +84,7 @@ public class HOME extends BetterLifeCommands implements CommandExecutor {
                     sendListOfHomes(player);
                     return false;
                 }
-                teleportPlayer(player, home, args[0]);
+                tpHelper.teleportPlayer(player, home, args[0]);
                 break;
             //TODO: Maybe we should add a /home info {home name} ?
             case 2:
@@ -108,85 +109,6 @@ public class HOME extends BetterLifeCommands implements CommandExecutor {
                 return false;
         }
         return true;
-    }
-
-    /**
-     * Teleports the given player to the location.
-     *
-     * @param player   Player to teleport.
-     * @param location Location to teleport player.
-     * @param home     The name of the home (used for command output).
-     */
-    private void teleportPlayer(Player player, Location location, String home) {
-        int ticks = ch.getCustomConfig().getInt("home-countdown") * 20;
-
-        sendMessage(player, "Teleporting to " + home + " in " + ticks / 20 + " seconds...", true);
-
-        final Location initial = player.getLocation();
-
-        //BL-Test
-        if(m.isTesting()) {
-            player.teleport(location);
-            return;
-        }
-
-        BukkitScheduler scheduler = Bukkit.getServer().getScheduler();
-        BukkitTask task = scheduler.runTaskLater(m, () -> {
-            player.teleport(location);
-            sendMessage(player, "&cTeleportation successful!", true);
-        }, ticks);
-
-        BukkitTask checker = scheduler.runTaskTimer(m, () -> {
-            if (!scheduler.isQueued(task.getTaskId())) {
-                cancelChecker();
-            }
-            Location cur = player.getLocation();
-
-            if (checkPlayerHasMoved(initial, cur)) {
-                if (scheduler.isQueued(task.getTaskId())) {
-                    cancelTask();
-                    cancelChecker();
-                    sendMessage(player, "&4Teleportation cancelled.", true);
-                }
-            }
-        }, 10, 10);
-
-        this.checkerID = checker.getTaskId();
-        this.taskID = task.getTaskId();
-    }
-
-    /**
-     * Cancels the teleport player checker.
-     */
-    private void cancelChecker() {
-        Bukkit.getServer().getScheduler().cancelTask(this.checkerID);
-    }
-
-    /**
-     * Cancels the current task.
-     */
-    private void cancelTask() {
-        Bukkit.getServer().getScheduler().cancelTask(this.taskID);
-    }
-
-    /**
-     * Checks whether or not the player has moved.
-     *
-     * @param initial Initial location of the player.
-     * @param cur     Current location of the player.
-     * @return True if the player has moved, false otherwise.
-     */
-    private boolean checkPlayerHasMoved(Location initial, Location cur) {
-        double maxX, maxZ, minX, minZ;
-
-        maxX = initial.getX() + 0.5;
-        maxZ = initial.getZ() + 0.5;
-
-        minX = initial.getX() - 0.5;
-        minZ = initial.getZ() - 0.5;
-
-        return !(cur.getX() >= minX && cur.getX() <= maxX
-                && cur.getZ() >= minZ && cur.getZ() <= maxZ);
     }
 
     /**
